requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.8.1
soundfile>=0.10.3
pyaudio>=0.2.11

# Speech processing
speechrecognition>=3.8.1
pydub>=0.25.1

# Speaker diarization
torch>=1.9.0
torchaudio>=0.9.0
speechbrain>=0.5.12
pyannote.audio>=2.1.1

# AI and NLP
requests>=2.25.1
openai>=0.27.0  # For potential future OpenAI integration
transformers>=4.21.0

# System integration
psutil>=5.8.0
AppKit>=0.2.8  # macOS app monitoring
Cocoa>=1.0     # macOS system integration

# Calendar integration
msal>=1.18.0   # Microsoft Graph authentication
microsoftgraph>=1.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.1
markdown>=3.3.4
pyyaml>=5.4.1
click>=8.0.1   # CLI interface

# Logging and monitoring
rich>=10.0.0   # Pretty terminal output
watchdog>=2.1.3  # File system monitoring
